refactor: 음식점 리스트를 컴포넌트로 분리
