+++ src/App.ts
@@ -1,11 +1,11 @@
 import Header from './components/Header';
-import AddModal from './components/AddModal';
+import Menu from './components/Menu';
 import FilterSection from './components/FilterSection';
-import RestaurantList from './components/RestaurantList';
+import RestaurantListSection from './components/RestaurantListSection';
+import AddModal from './components/AddModal';
 import RestaurantListItem from './domain/RestaurantListItem';
 import RestaurantStorage from './domain/RestaurantStorage';
 import { ID } from './constants';
-import Menu from './components/Menu';
 
 class App {
   #app;
@@ -23,7 +23,7 @@ class App {
       <main>
         ${Menu.template()}
         ${FilterSection.template()}
-        ${RestaurantList.template(this.#restaurantListItem.getListItem())}
+        ${RestaurantListSection.template(this.#restaurantListItem.getListItem())}
 
         ${AddModal.template()}
       </main>`;
@@ -35,7 +35,7 @@ class App {
     Header.setEvent();
     Menu.setEvent(this.#restaurantListItem);
     FilterSection.setEvent(this.#restaurantListItem);
-    RestaurantList.setEvent(this.#restaurantListItem);
+    RestaurantListSection.setEvent(this.#restaurantListItem);
     AddModal.setEvent(this.#restaurantListItem);
   }
 }
