+++ src/components/RestaurantListSection/FavoriteIcon.ts
