+++ src/components/RestaurantListSection/index.ts
@@ -1,14 +1,13 @@
 import { CLASS } from '../../constants';
 import RestaurantListItem, { IRestaurant } from '../../domain/RestaurantListItem';
 import Restaurant from './Restaurant';
+import RestaurantList from './RestaurantList';
 
-const RestaurantList = {
+const RestaurantListSection = {
   template(restaurantList: IRestaurant[]) {
     return `
     <section class="${CLASS.RESTAURANT_LIST_CONTAINER}">
-      <ul class='restaurant-list'>
-        ${restaurantList.map((restaurant) => Restaurant.template(restaurant)).join('')}
-      </ul>
+      ${RestaurantList.template(restaurantList)}
     </section>`;
   },
   setEvent(RestaurantListItem: RestaurantListItem) {
@@ -16,7 +15,7 @@ const RestaurantList = {
   },
   update(RestaurantListItem: RestaurantListItem, restaurantList: IRestaurant[]) {
     const restaurantListContainer = document.querySelector(`.${CLASS.RESTAURANT_LIST_CONTAINER}`) as HTMLDivElement;
-    restaurantListContainer.innerHTML = this.template(restaurantList);
+    restaurantListContainer.innerHTML = RestaurantList.template(restaurantList);
     this.setEvent(RestaurantListItem);
   },
   append(RestaurantListItem: RestaurantListItem, restaurant: IRestaurant) {
@@ -27,4 +26,4 @@ const RestaurantList = {
   },
 };
 
-export default RestaurantList;
+export default RestaurantListSection;
